requirements.txt
# Core data processing libraries
pandas~=2.2.0
pyarrow~=19.0.0
duckdb~=1.2.0
fastparquet==2024.11.0
numpy~=2.2.0

# Visualization libraries
matplotlib~=3.10.0
seaborn~=0.13.0
plotly~=6.0.0
dash~=3.0.0
dash-bootstrap-components~=2.0.0

# Machine learning
scikit-learn~=1.6.0

# Utilities
python-dotenv~=1.1.0
tqdm~=4.67.0

# Development and notebook support
ipykernel~=6.29.0
jupyterlab~=4.4.0

# Data transformation
dbt-duckdb~=1.9.0

# Vanna.ai for natural language to SQL
vanna[chromadb,gemini,duckdb]~=0.7.0
vertexai~=1.71.0
